Refactor: Extract List Table to a separated component
