+++ src/EventTickets/resources/admin/components/EventDetailsPage/TicketTypesSection/TicketTypesSection.module.scss
@@ -1,23 +1,3 @@
-.ticketTypesSection {
-    table {
-        tr {
-            th, td {
-                padding-inline: var(--givewp-spacing-9);
-
-                &:first-child {
-                    display: none;
-                }
-            }
-        }
-
-        thead {
-            th {
-                color: var(--givewp-grey-700);
-            }
-        }
-    }
-}
-
 .container {
     text-align: center;
     color: #424242;
